refactor: Separate class Contribution from AdminOverlay file; Correct formatting
